services/api/requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# OpenSearch client
opensearch-py==2.4.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Database
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1

# Redis
redis==5.0.1

# Logging and monitoring
structlog==24.1.0
python-json-logger==2.0.7
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
httpx==0.26.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.26.0
